+++ packages/launchpad/src/setup/InstallDependencies.cy.tsx
@@ -15,9 +15,13 @@ describe('<InstallDependencies />', () => {
       },
     })
 
-    cy.contains('a', 'react-scripts')
+    cy.contains('a', 'react')
     .should('be.visible')
-    .and('have.attr', 'href', 'https://www.npmjs.com/package/react-scripts')
+    .and('have.attr', 'href', 'https://www.npmjs.com/package/react')
+
+    cy.contains('a', 'react-dom')
+    .should('be.visible')
+    .and('have.attr', 'href', 'https://www.npmjs.com/package/react-dom')
 
     cy.contains('a', 'typescript')
     .should('be.visible')
@@ -65,10 +69,17 @@ describe('<InstallDependencies />', () => {
         gqlVal.wizard.packagesToInstall = [
           {
             __typename: 'WizardNpmPackage',
-            id: 'cra',
+            id: 'react',
             satisfied: true,
-            detectedVersion: '1.0.1',
-            ...wizardDeps.WIZARD_DEPENDENCY_REACT_SCRIPTS,
+            detectedVersion: '18.3.1',
+            ...wizardDeps.WIZARD_DEPENDENCY_REACT,
+          },
+          {
+            __typename: 'WizardNpmPackage',
+            id: 'react-dom',
+            satisfied: true,
+            detectedVersion: '18.3.1',
+            ...wizardDeps.WIZARD_DEPENDENCY_REACT_DOM,
           },
           {
             __typename: 'WizardNpmPackage',