+++ system-tests/projects/invalid-root-level-config/invalid-component-experimentalSessionAndOrigin-config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  component: {
-    devServer: {
-      framework: 'create-react-app',
-      bundler: 'webpack',
-    },
-  },
-}