+++ system-tests/projects/pristine/expected-cypress-ts-component-create-react-app-v5-without-fixtures/cypress.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'cypress'
-import { devServer } from '@cypress/react/plugins/react-scripts'
-
-export default defineConfig({
-  component: {
-    devServer,
-  },
-})