+++ system-tests/test/component_testing_spec.ts
@@ -3,14 +3,6 @@ import systemTests from '../lib/system-tests'
 describe('component testing projects', function () {
   systemTests.setup()
 
-  systemTests.it('create-react-app-configured', {
-    project: 'create-react-app-configured',
-    testingType: 'component',
-    spec: 'src/App.cy.js',
-    browser: 'chrome',
-    expectedExitCode: 0,
-  })
-
   systemTests.it('vueclivue3-configured', {
     project: 'vueclivue3-configured',
     testingType: 'component',